# GIF export speedup