I looked into LZW compression code in the gifexporter addon and simplified it a little bit.

The result is as speedup of around 3.6 times.

Using Roroto's awesome splashscreen it went from 43,26 seconds to 11,96 seconds in the editor. Compiled version of Pixelorama goes from 34.5 to 9.3 seconds.

I did some tests with multiple animation files and gifs exported from 0.10 and exported from this changed version are identical according to "cmp -l"

I will eventually commit this also to the jegor's repository but it seems to be outdated at this moment.

I also commented the decompress function because it now needs a rewrite and is currently unused.

There are for sure more ways the speed the whole thing up but small steps...