@novhack 2022-05-07
Do not merge yet. I found a bug with big images which require quantization. They are exported as solid color gifs and I need to investigate.