@novhack 2022-05-08
I rewrote the add_image_data_block() function to have less looping. It shaves of only around 100ms from the export I test on but it hugely helps in a profiler where it was causing big spikes.