@novhack 2022-05-08
I am done with this PR for now. The next step would be to somehow speed up LSBLZWBitPacker.write_bits() function which gets called a lot but it already uses a lot of bithacks so there is not much room to work with.