@novhack 2022-05-12
The last commit helped to shave off a few hundred milliseconds. Now the export takes around 11.5 seconds.
The next step is to add bulk export - allow multiple frames to be processed at the same time in a thread pool.